Remove controller of task gantt
