+++ app/Http/Controllers/Task/TaskController.php
@@ -20,6 +20,7 @@ use Jitamin\Filter\TaskProjectFilter;
 use Jitamin\Filter\TaskProjectsFilter;
 use Jitamin\Filter\TaskTitleFilter;
 use Jitamin\Formatter\TaskAutoCompleteFormatter;
+use Jitamin\Formatter\TaskGanttFormatter;
 use Jitamin\Model\TaskModel;
 use Jitamin\Model\UserMetadataModel;
 
@@ -76,6 +77,31 @@ class TaskController extends BaseController
             'external_links'  => $this->taskExternalLinkModel->getAll($task['id']),
             'link_label_list' => $this->linkModel->getList(0, false),
             'tags'            => $this->taskTagModel->getList($task['id']),
+        ]));
+    }
+
+    /**
+     * Show Gantt chart for one project.
+     */
+    public function gantt()
+    {
+        $project = $this->getProject();
+        $search = $this->helper->projectHeader->getSearchQuery($project);
+        $sorting = $this->request->getStringParam('sorting', 'board');
+        $filter = $this->taskLexer->build($search)->withFilter(new TaskProjectFilter($project['id']));
+
+        if ($sorting === 'date') {
+            $filter->getQuery()->asc(TaskModel::TABLE.'.date_started')->asc(TaskModel::TABLE.'.date_creation');
+        } else {
+            $filter->getQuery()->asc('column_position')->asc(TaskModel::TABLE.'.position');
+        }
+
+        $this->response->html($this->helper->layout->app('task/gantt', [
+            'project'     => $project,
+            'title'       => $project['name'],
+            'description' => $this->helper->projectHeader->getDescription($project),
+            'sorting'     => $sorting,
+            'tasks'       => $filter->format(new TaskGanttFormatter($this->container)),
         ]));
     }
 
@@ -233,6 +259,27 @@ class TaskController extends BaseController
         $task = $this->getTask();
         $this->taskModel->update(['id' => $task['id'], 'date_started' => time()]);
         $this->response->redirect($this->helper->url->to('Task/TaskController', 'show', ['project_id' => $task['project_id'], 'task_id' => $task['id']]));
+    }
+
+    /**
+     * Save new task start date and due date.
+     */
+    public function set_date()
+    {
+        $this->getProject();
+        $values = $this->request->getJson();
+
+        $result = $this->taskModel->update([
+            'id'           => $values['id'],
+            'date_started' => strtotime($values['start']),
+            'date_due'     => strtotime($values['end']),
+        ]);
+
+        if (!$result) {
+            $this->response->json(['message' => 'Unable to save task'], 400);
+        } else {
+            $this->response->json(['message' => 'OK'], 201);
+        }
     }
 
     /**
