+++ resources/views/project_user_overview/roles.php
@@ -14,7 +14,7 @@
             </td>
             <td>
                 <?= $this->url->link('<i class="fa fa-columns"></i>', 'Project/Board/BoardController', 'show', ['project_id' => $project['id']], false, 'dashboard-table-link', t('Board')) ?>
-                <?= $this->url->link('<i class="fa fa-sliders"></i>', 'Task/TaskGanttController', 'show', ['project_id' => $project['id']], false, 'dashboard-table-link', t('Gantt chart')) ?>
+                <?= $this->url->link('<i class="fa fa-sliders"></i>', 'Task/TaskController', 'gantt', ['project_id' => $project['id']], false, 'dashboard-table-link', t('Gantt chart')) ?>
                 <?= $this->url->link('<i class="fa fa-cog"></i>', 'Project/ProjectSettingsController', 'show', ['project_id' => $project['id']], false, 'dashboard-table-link', t('Project settings')) ?>
 
                 <?= $this->text->e($project['project_name']) ?>
